game_project/requirements.txt
Pillow>=10.0.0
tkinter  # Usually comes with Python installation 